Added a shim to convert between 0- and 1-based values
Previously we did this in the Puzzle class. That proved to be the wrong
boundary. Doing the conversion there meant that many tests had to deal
with a mix of 0- and 1-based values. Pushing that to a shim between the
rest of the system and the external sudoku module means that the conversion
is now an implementation detail and can be ignored by almost all the tests.